src/omop_lib/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "omop-lib"
version = "0.1.0"
description = "Shared query + diagnostics helpers for the OMOP Explorer apps."
requires-python = ">=3.10"
dependencies = [
  "databricks-sql-connector==3.7.0",
  "databricks-sdk==0.38.0",
]

[tool.setuptools]
packages = ["omop_lib"]
